+++ packages/picasso.js/src/about.js
@@ -1,1 +1,1 @@
-export default { version: '2.3.0' };
+export default { version: '2.3.1' };